Moved loadstore to array_utils
